Isolate write ops for readability
